feat: Improve loading
